Change type in column

+++ database/migrations/2022_02_22_075623_create_consultation_category_table.php
@@ -16,7 +16,7 @@ class CreateConsultationCategoryTable extends Migration
         Schema::create('category_consultation', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('consultation_id');
-            $table->unsignedInteger('category_id');
+            $table->unsignedBigInteger('category_id');
             $table->timestamps();
 
             $table->foreign('consultation_id')->on('consultations')->references('id')->cascadeOnDelete();
